requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
gunicorn==19.9.0
Markdown==3.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2018.9
regex==2019.3.12
static3==0.7.0
